Training module restructuring
